Blam cleanup part 3
TODO: himanshu01, Reflexive fix for tag injector
