--- xlive/Blam/BlamLibrary.h
@@ -1,4 +1,0 @@
-#pragma once
-#include "Engine\EngineDefinitions.h"
-#include "Enums\Enums.h"
-#include "Maths\Maths.h"
