+++ xlive/Blam/Cache/Tags/itmc.cpp
@@ -1,20 +1,20 @@
-#include"itmc.h"
+#include "itmc.h"
 
-namespace Blam
-{
-	namespace Cache
-	{
-		namespace Tags
-		{
-			void itmc::Open()
-			{
-				ItemPermutations.Open();
-			}
-			void itmc::Close()
-			{
-				ItemPermutations.Close();
-			}
-
-		}
-	}
-}
+// TODO: Himanshu01 fix this pls
+//namespace Blam
+//{
+//	namespace Cache
+//	{
+//		namespace Tags
+//		{
+//			void itmc::Open()
+//			{
+//				ItemPermutations.Open();
+//			}
+//			void itmc::Close()
+//			{
+//				ItemPermutations.Close();
+//			}
+//		}
+//	}
+//}
