+++ xlive/Blam/Cache/Tags/itmc.h
@@ -1,5 +1,7 @@
 #pragma once
+
 #include "tag_base.h"
+#include "Blam/Reflexive/Reflexive.h"
 
 namespace Blam
 {
@@ -9,14 +11,14 @@ namespace Blam
 		{
 			struct itmc : tag_base
 			{
-				struct ItemPermutations :tag_base
+				struct ItemPermutations : tag_base
 				{
 					float Weight;
 					tag_reference Item;
 					string_id VariantName;
 				};
 				TAG_BLOCK_SIZE_ASSERT(ItemPermutations, 0x10);
-				Blam::Cache::DataTypes::Reflexive<ItemPermutations> ItemPermutations;
+				Blam::Cache::DataTypes::Reflexive<ItemPermutations> itemPermutations;
 				__int32 UnusedSpawnTime;
 
 				//override
