+++ xlive/Blam/Cache/Tags/scnr.cpp
@@ -1,19 +1,20 @@
 #include "scnr.h"
 
-namespace Blam
-{
-	namespace Cache
-	{
-		namespace Tags
-		{
-			void scnr::Open()
-			{
-				SimulationDefinitionTable.Open();
-			}
-			void scnr::Close()
-			{
-				SimulationDefinitionTable.Close();
-			}
-		}
-	}
-}
+// TODO: Himanshu01 pls fix
+//namespace Blam
+//{
+//	namespace Cache
+//	{
+//		namespace Tags
+//		{
+//			void scnr::Open()
+//			{
+//				SimulationDefinitionTable.Open();
+//			}
+//			void scnr::Close()
+//			{
+//				SimulationDefinitionTable.Close();
+//			}
+//		}
+//	}
+//}
