+++ xlive/Blam/Cache/Tags/tag_definitons.h
@@ -1,4 +1,5 @@
 #pragma once
 
-#include"itmc.h"
-#include"scnr.h"
+// TODO: himanshu01 please fix
+//#include "itmc.h"
+//#include "scnr.h"
