+++ xlive/Blam/Common/Common.h
@@ -1,0 +1,10 @@
+#pragma once
+
+enum scnr_type : int
+{
+	SinglePlayer = 1,
+	Multiplayer,
+	MainMenu,
+	MultiplayerShared,
+	SinglePlayerShared
+};
