+++ xlive/Blam/Engine/EngineDefinitions.h
@@ -1,11 +1,7 @@
 #pragma once
-#include "stdafx.h"
-#include "Players\Players.h"
 
 #include "Game\GameManager.h"
 #include "Game\GameTimeGlobals.h"
 #include "Game\HaloScript.h"
 
-#include "Objects\Objects.h"
-#include "Objects\ObjectPlacementData.h"
 #include "Actor\Actor.h"
