+++ xlive/Blam/Enums/Enums.h
@@ -1,7 +1,4 @@
 #pragma once
-#include "stdafx.h"
-#include "Players\Players.h"
+
 #include "Game\GameTimeGlobals.h"
 #include "Game\HaloScriptTypes.h"
-#include "Objects\Objects.h"
-#include "Objects\Unit.h"
