+++ xlive/Blam/Enums/Game/HaloScriptTypes.h
@@ -1,5 +1,4 @@
 #pragma once
-#include <unordered_map>
 
 namespace Blam
 {
