+++ xlive/H2MOD/Variants/VariantMPGameEngine.cpp
@@ -1,4 +1,4 @@
-#include "stdafx.h"
+
 #include "VariantMPGameEngine.h"
 #include "H2MOD\Modules\Startup\Startup.h"
 
@@ -255,7 +255,7 @@ c_engine_func_proto(unk_function_27_wrapper, int)
 }
 c_engine_func_proto(is_team_enemy_wrapper, int, BYTE team_a, BYTE team_b)
 {
-	return GET_ENGINE()->is_team_enemy(static_cast<Team>(team_a), static_cast<Team>(team_b));
+	return GET_ENGINE()->is_team_enemy(static_cast<ObjectTeam>(team_a), static_cast<ObjectTeam>(team_b));
 }
 c_engine_func_proto(unk_function_29_wrapper, void, int arg1)
 {
@@ -465,7 +465,7 @@ int c_game_engine_base::unk_function_27()
 {
 	return call_c_base_func(unk_function_27);
 }
-bool c_game_engine_base::is_team_enemy(Team team_a, Team team_b)
+bool c_game_engine_base::is_team_enemy(ObjectTeam team_a, ObjectTeam team_b)
 {
 	return call_c_base_func(is_team_enemy, team_a, team_b);
 }
