+++ xlive/H2MOD/Variants/VariantMPGameEngine.h
@@ -1,7 +1,6 @@
 #pragma once
-#include "Blam\Enums\Players\Players.h"
 
-using namespace Blam::Enums::Player;
+#include "Blam\Engine\Objects\Objects.h"
 
 enum c_game_engine_types
 {
@@ -62,7 +61,7 @@ public:
 	virtual int unk_function_25(int arg1, int arg2);
 	virtual int unk_function_26(signed int arg1, int arg2, signed int arg3);
 	virtual int unk_function_27();
-	virtual bool is_team_enemy(Team team_a, Team team_b);
+	virtual bool is_team_enemy(ObjectTeam team_a, ObjectTeam team_b);
 	virtual void unk_function_29(int arg1);
 	virtual void unk_function_30(signed int arg1, int arg2, signed int arg3);
 	virtual void unk_function_31(int arg1, int arg2, char arg3, signed int arg4);
